orders/requirements.txt
aiofiles==22.1.0
aiosqlite==0.18.0
aniso8601==9.0.1
asyncio==3.4.3
asyncpg==0.27.0
click==8.1.3
Flask==2.2.2
Flask-RESTful==0.3.9
greenlet==2.0.1
httptools==0.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.19.0
multidict==6.0.4
nanoid==2.0.0
nats-py==2.2.0
packaging==22.0
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.7
sanic==22.9.1
Sanic-Cors==2.2.0
sanic-routing==22.8.0
six==1.16.0
SQLAlchemy==1.4.45
typing_extensions==4.4.0
ujson==5.6.0
uvloop==0.17.0
websockets==10.4
Werkzeug==2.2.2
zone-common==0.0.17
